refactor(ImageAreaSelectorFeature): rename selector components
