+++ src/features/image-area-selector/components/ImageAreaSelector/components/SelectionArea.tsx
