+++ src/features/image-area-selector/components/ImageAreaSelector/constants/cursor.ts
