+++ src/features/image-area-selector/components/ImageAreaSelector/constants/position.ts
