+++ src/features/image-area-selector/components/ImageAreaSelector/hooks/cursor.ts
