+++ src/features/image-area-selector/components/ImageAreaSelector/hooks/position.ts
