+++ src/features/image-area-selector/components/ImageAreaSelector/utils.ts
